redis: refactoring to use the new pattern
